requirements.txt
# Computer Vision
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.3.0

# NLP
transformers>=4.20.0
pytesseract>=0.3.8

# PDF Processing
pdf2image>=1.16.0
PyPDF2>=2.10.0

# Utilitaires
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0

# Interface (optionnel)
streamlit>=1.10.0
Flask>=2.0.0

# Logging et configuration
python-dotenv>=0.19.0
pyyaml>=5.4.0

# Tests
pytest>=6.2.0
pytest-cov>=2.12.0

sentencepiece>=0.2.1